Revert "fix: button layouts"
