+++ package.json
@@ -17,7 +17,6 @@
     "react-dom": "^16.13.1",
     "react-html-parser": "^2.0.2",
     "react-scripts": "3.4.1",
-    "react-select": "^3.1.0",
     "react-table-6": "^6.11.0",
     "typescript": "^3.9.7"
   },
